First attempt to use SLURM srun.
In a srun command we lose the context of the container on a remote node,
we need to srun a new container and start the process with given parameters.

Using a file in the container that launches the srun command won't work,
as the container context does not exist on the remote node.